backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
supabase==2.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
httpx==0.27.2
joblib==1.4.2
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.2
python-multipart==0.0.9
